# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jet on Set API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/profile/account": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Display current User data.",
                "description": "User data Page",
                "operationId": "3b7669a21106dcc0fc95a2afb8ed5620",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Update user data",
                "operationId": "b608ed5c0da964d36bbbc3b5a36fa182",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "ihamzehald@gmail.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "larapoints123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Destroy the specified resource in storage.",
                "description": "User destroy",
                "operationId": "4876eaad1de6e7acb6100a0d448d71f7",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/profile/account/login": {
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "User login",
                "operationId": "da5d84ba5deb0f988bb6de86be0d600a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "ihamzehald@gmail.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "larapoints123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/profile/account/register": {
            "post": {
                "tags": [
                    "Account"
                ],
                "summary": "Create a new user instance after a valid registration.",
                "description": "User register",
                "operationId": "56477b20e88036a71c9df41a9644dd86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "First Name",
                                        "type": "string",
                                        "example": "first name"
                                    },
                                    "last_name": {
                                        "description": "Last Name",
                                        "type": "string",
                                        "example": "last name"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "ihamzehald@gmail.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "larapoints123"
                                    },
                                    "password_confirmation": {
                                        "description": "User password confirmation",
                                        "type": "string",
                                        "example": "larapoints123"
                                    },
                                    "privacy": {
                                        "description": "Privacy",
                                        "type": "boolean",
                                        "example": "true"
                                    },
                                    "terms": {
                                        "description": "Terms",
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/profile/account/refresh": {
            "put": {
                "tags": [
                    "Account"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "User refresh",
                "operationId": "f9b3fcc30dfb26e056ac8270478f99ed",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}